+++ package.json
@@ -14,29 +14,30 @@
     "debug": "tauri build -- --debug"
   },
   "dependencies": {
-    "@fortawesome/free-solid-svg-icons": "^6.5.2",
-    "@svelteuidev/composables": "^0.14.0",
-    "@tauri-apps/api": "^2.0.0",
-    "@tauri-apps/plugin-clipboard-manager": "^2.2.0",
-    "@tauri-apps/plugin-dialog": "^2.0.0",
-    "@tauri-apps/plugin-shell": "^2.0.0",
+    "@fortawesome/free-solid-svg-icons": "^7.0.0",
+    "@svelteuidev/composables": "^0.15.7",
+    "@tauri-apps/api": "^2.7.0",
+    "@tauri-apps/plugin-clipboard-manager": "^2.3.0",
+    "@tauri-apps/plugin-dialog": "^2.3.1",
+    "@tauri-apps/plugin-shell": "^2.3.0",
     "dayjs": "^1.11.13",
-    "mathjs": "^11.10.0",
-    "svelte-fa": "^4.0.2",
-    "svelte-virtuallists": "^1.4.1",
+    "mathjs": "^14.5.3",
+    "svelte-fa": "^4.0.4",
+    "svelte-virtuallists": "^1.4.2",
     "tauri-plugin-store-api": "github:tauri-apps/tauri-plugin-store"
   },
   "devDependencies": {
-    "@sveltejs/adapter-static": "^3.0.0",
-    "@sveltejs/kit": "^2.21.0",
-    "@sveltejs/vite-plugin-svelte": "^5.0.3",
-    "@tauri-apps/cli": "^2.0.2",
-    "sass": "1.60.0",
+    "@sveltejs/adapter-static": "^3.0.8",
+    "@sveltejs/kit": "^2.25.2",
+    "@sveltejs/vite-plugin-svelte": "^6.1.0",
+    "@tauri-apps/cli": "^2.7.1",
+    "sass": "1.89.2",
     "simple-svelte-autocomplete": "^2.5.2",
-    "svelte": "^5.0.0",
-    "svelte-check": "^4.0.0",
-    "svelte-preprocess": "^6.0.0",
-    "typescript": "^5.5.0",
-    "vite": "^6.3.5"
-  }
+    "svelte": "^5.36.15",
+    "svelte-check": "^4.3.0",
+    "svelte-preprocess": "^6.0.3",
+    "typescript": "^5.8.3",
+    "vite": "^7.0.6"
+  },
+  "packageManager": "pnpm@9.10.0+sha512.73a29afa36a0d092ece5271de5177ecbf8318d454ecd701343131b8ebc0c1a91c487da46ab77c8e596d6acf1461e3594ced4becedf8921b074fbd8653ed7051c"
 }
