+++ src-tauri/.editorconfig
@@ -1,0 +1,3 @@
+[*]
+indent_size = 4
+indent_style = space
