+++ src-tauri/Cargo.toml
@@ -10,7 +10,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [build-dependencies]
-tauri-build = { version = "2", features = [] }
+tauri-build = { version = "2.3.1", features = [] }
 
 [dependencies]
 opener = { version = "0.7.2", featuers = ["reveal"] }
@@ -18,7 +18,6 @@ serde_json = "1.0"
 chrono = "0.4.24"
 regex = "1"
 serde = { version = "1.0", features = ["derive"] }
-tauri = { version = "2", features = ["protocol-asset"] }
 vdm = "1.0.9"
 reqwest = { version = "0.11", features = ["blocking", "json"] }
 steamid-ng = "1.0.0"
@@ -29,13 +28,14 @@ fnv = "1.0.7"
 bitbuffer = { version = "0.11.0", features = ["serde"] }
 num_enum = "0.5.7"
 parse-display = "0.8.0"
-tauri-plugin-store = { version = "2" }
-tauri-plugin-dialog = "2"
-tauri-plugin-shell = "2"
 sanitize-filename = "0.6.0"
-tauri-plugin-clipboard-manager = "2.2.0"
 trash = "3"
 fs_extra = "1.3.0"
+tauri = { version = "2.7.0", features = ["protocol-asset"] }
+tauri-plugin-store = "2.3.0"
+tauri-plugin-dialog = "2.3.1"
+tauri-plugin-shell = "2.3.0"
+tauri-plugin-clipboard-manager = "2.3.0"
 
 [features]
 # by default Tauri runs in production mode
