+++ src-tauri/gen/schemas/windows-schema.json
@@ -21,9 +21,7 @@
     {
       "description": "A list of capabilities.",
       "type": "object",
-      "required": [
-        "capabilities"
-      ],
+      "required": ["capabilities"],
       "properties": {
         "capabilities": {
           "description": "The list of capabilities.",
@@ -39,10 +37,7 @@
     "Capability": {
       "description": "A grouping and boundary mechanism developers can use to isolate access to the IPC layer.\n\nIt controls application windows fine grained access to the Tauri core, application, or plugin commands. If a window is not matching any capability then it has no access to the IPC layer at all.\n\nThis can be done to create groups of windows, based on their required system access, which can reduce impact of frontend vulnerabilities in less privileged windows. Windows can be added to a capability by exact name (e.g. `main-window`) or glob patterns like `*` or `admin-*`. A Window can have none, one, or multiple associated capabilities.\n\n## Example\n\n```json { \"identifier\": \"main-user-files-write\", \"description\": \"This capability allows the `main` window on macOS and Windows access to `filesystem` write related commands and `dialog` commands to enable programatic access to files selected by the user.\", \"windows\": [ \"main\" ], \"permissions\": [ \"core:default\", \"dialog:open\", { \"identifier\": \"fs:allow-write-text-file\", \"allow\": [{ \"path\": \"$HOME/test.txt\" }] }, \"platforms\": [\"macOS\",\"windows\"] } ```",
       "type": "object",
-      "required": [
-        "identifier",
-        "permissions"
-      ],
+      "required": ["identifier", "permissions"],
       "properties": {
         "identifier": {
           "description": "Identifier of the capability.\n\n## Example\n\n`main-user-files-write`",
@@ -93,10 +88,7 @@
         },
         "platforms": {
           "description": "Limit which target platforms this capability applies to.\n\nBy default all platforms are targeted.\n\n## Example\n\n`[\"macOS\",\"windows\"]`",
-          "type": [
-            "array",
-            "null"
-          ],
+          "type": ["array", "null"],
           "items": {
             "$ref": "#/definitions/Target"
           }
@@ -106,9 +98,7 @@
     "CapabilityRemote": {
       "description": "Configuration for remote URLs that are associated with the capability.",
       "type": "object",
-      "required": [
-        "urls"
-      ],
+      "required": ["urls"],
       "properties": {
         "urls": {
           "description": "Remote domains this capability refers to using the [URLPattern standard](https://urlpattern.spec.whatwg.org/).\n\n## Examples\n\n- \"https://*.mydomain.dev\": allows subdomains of mydomain.dev - \"https://mydomain.dev/api/*\": allows any subpath of mydomain.dev/api",
@@ -207,10 +197,7 @@
                       "anyOf": [
                         {
                           "type": "object",
-                          "required": [
-                            "cmd",
-                            "name"
-                          ],
+                          "required": ["cmd", "name"],
                           "properties": {
                             "args": {
                               "description": "The allowed arguments for the command execution.",
@@ -233,10 +220,7 @@
                         },
                         {
                           "type": "object",
-                          "required": [
-                            "name",
-                            "sidecar"
-                          ],
+                          "required": ["name", "sidecar"],
                           "properties": {
                             "args": {
                               "description": "The allowed arguments for the command execution.",
@@ -267,10 +251,7 @@
                       "anyOf": [
                         {
                           "type": "object",
-                          "required": [
-                            "cmd",
-                            "name"
-                          ],
+                          "required": ["cmd", "name"],
                           "properties": {
                             "args": {
                               "description": "The allowed arguments for the command execution.",
@@ -293,10 +274,7 @@
                         },
                         {
                           "type": "object",
-                          "required": [
-                            "name",
-                            "sidecar"
-                          ],
+                          "required": ["name", "sidecar"],
                           "properties": {
                             "args": {
                               "description": "The allowed arguments for the command execution.",
@@ -345,20 +323,14 @@
                 },
                 "allow": {
                   "description": "Data that defines what is allowed by the scope.",
-                  "type": [
-                    "array",
-                    "null"
-                  ],
+                  "type": ["array", "null"],
                   "items": {
                     "$ref": "#/definitions/Value"
                   }
                 },
                 "deny": {
                   "description": "Data that defines what is denied by the scope. This should be prioritized by validation logic.",
-                  "type": [
-                    "array",
-                    "null"
-                  ],
+                  "type": ["array", "null"],
                   "items": {
                     "$ref": "#/definitions/Value"
                   }
@@ -366,9 +338,7 @@
               }
             }
           ],
-          "required": [
-            "identifier"
-          ]
+          "required": ["identifier"]
         }
       ]
     },
@@ -2227,37 +2197,27 @@
         {
           "description": "MacOS.",
           "type": "string",
-          "enum": [
-            "macOS"
-          ]
+          "enum": ["macOS"]
         },
         {
           "description": "Windows.",
           "type": "string",
-          "enum": [
-            "windows"
-          ]
+          "enum": ["windows"]
         },
         {
           "description": "Linux.",
           "type": "string",
-          "enum": [
-            "linux"
-          ]
+          "enum": ["linux"]
         },
         {
           "description": "Android.",
           "type": "string",
-          "enum": [
-            "android"
-          ]
+          "enum": ["android"]
         },
         {
           "description": "iOS.",
           "type": "string",
-          "enum": [
-            "iOS"
-          ]
+          "enum": ["iOS"]
         }
       ]
     },
@@ -2271,9 +2231,7 @@
         {
           "description": "A variable that is set while calling the command from the webview API.",
           "type": "object",
-          "required": [
-            "validator"
-          ],
+          "required": ["validator"],
           "properties": {
             "raw": {
               "description": "Marks the validator as a raw regex, meaning the plugin should not make any modification at runtime.\n\nThis means the regex will not match on the entire string by default, which might be exploited if your regex allow unexpected input to be considered valid. When using this option, make sure your regex is correct.",
