+++ src-tauri/src/main.rs
@@ -8,15 +8,16 @@ use chrono::prelude::*;
 use regex::Regex;
 use sanitize_filename::sanitize;
 use serde_json::{self, json, Value};
+use std::env::{current_dir, home_dir};
+use std::ffi::OsStr;
 use std::fs::{DirEntry, File};
 use std::io::Write;
 use std::path::{Path, PathBuf};
-use std::process::Command;
+use std::sync::LazyLock;
 use std::{env, fs};
 use tauri::command;
 use vdm::action::ActionType;
 use vdm::VDM;
-use trash;
 
 use crate::clip::Clip;
 use crate::demos::*;
@@ -33,8 +34,8 @@ mod event;
 mod logstf;
 mod macros;
 mod settings;
-mod vdms;
 mod tf2;
+mod vdms;
 
 fn setting_as_bin(setting: &Value) -> i64 {
     if !setting.is_boolean() {
@@ -74,59 +75,59 @@ fn write_cfg(settings: &Value) -> Result<(), String> {
 
     extend!(
         cfg,
-        "echo \"Execing Melies Config\";\r\ncl_drawhud {};\r\n",
+        "echo \"Execing Melies Config\";\ncl_drawhud {};\n",
         setting_as_bin(&settings["output"]["hud"])
     );
-    extend!(cfg, "sv_cheats {};\r\n", "1");
+    extend!(cfg, "sv_cheats {};\n", "1");
     extend!(
         cfg,
-        "voice_enable {};\r\n",
+        "voice_enable {};\n",
         setting_as_bin(&settings["output"]["voice_chat"])
     );
     extend!(
         cfg,
-        "hud_saytext_time {};\r\n",
+        "hud_saytext_time {};\n",
         setting_as_bin(&settings["output"]["text_chat"]) * 12
     );
     extend!(
         cfg,
-        "crosshair {};\r\n",
+        "crosshair {};\n",
         setting_as_bin(&settings["output"]["crosshair"])
     );
     extend!(
         cfg,
-        "r_drawviewmodel {};\r\n",
+        "r_drawviewmodel {};\n",
         setting_as_bin(&settings["output"]["viewmodel"])
     );
     extend!(
         cfg,
-        "tf_use_min_viewmodels {};\r\n",
+        "tf_use_min_viewmodels {};\n",
         setting_as_bin(&settings["output"]["minmode"])
     );
     extend!(
         cfg,
-        "viewmodel_fov_demo {};\r\n",
+        "viewmodel_fov_demo {};\n",
         setting_as_bin(&settings["recording"]["viewmodel_fov"])
     );
     extend!(
         cfg,
-        "fov_desired {};\r\n",
+        "fov_desired {};\n",
         setting_as_bin(&settings["recording"]["fov"])
     );
 
     if setting_as_bin(&settings["recording"]["third_person"]) == 1 {
-        extend!(cfg, "thirdperson{};\r\n", "");
+        extend!(cfg, "thirdperson{};\n", "");
     } else {
-        extend!(cfg, "firstperson{};\r\n", "");
+        extend!(cfg, "firstperson{};\n", "");
     }
 
     if let Some(commands) = settings["recording"]["commands"].as_str() {
-        extend!(cfg, "{}\r\n", commands);
+        extend!(cfg, "{}\n", commands);
     }
 
     extend!(
         cfg,
-        "{};\r\n",
+        "{};\n",
         "alias \"snd_fix\" \"snd_restart; snd_soundmixer Default_mix;\""
     );
 
@@ -138,9 +139,8 @@ fn write_cfg(settings: &Value) -> Result<(), String> {
     };
 
     let base_folder = Path::new(tf_folder).parent().unwrap();
-    let base_folder_str = base_folder.to_str().unwrap();
 
-    let mut installs: Vec<String> = vec!["tf".to_string()];
+    let mut installs = vec![PathBuf::from("tf")];
 
     for install in settings["alt_installs"].as_array().unwrap() {
         if let None = install["name"].as_str() {
@@ -151,19 +151,17 @@ fn write_cfg(settings: &Value) -> Result<(), String> {
             continue;
         }
 
-        let mut install_folder = install["tf_folder"].as_str().unwrap().replace(base_folder_str, "");
+        let install_folder = Path::new(install["tf_folder"].as_str().unwrap())
+            .strip_prefix(base_folder)
+            .unwrap();
 
-        if install_folder.starts_with("\\") {
-            install_folder.remove(0);
-        }
-
-        installs.push(install_folder);
+        installs.push(install_folder.to_owned());
     }
 
     for install in installs {
-        let install_folder_str = format!("{}\\{}", base_folder_str, install);
+        let install_folder = base_folder.join(install);
 
-        let cfg_path = format!("{}\\cfg", install_folder_str);
+        let cfg_path = install_folder.join("cfg");
 
         if !Path::new(&cfg_path).exists() {
             match std::fs::create_dir_all(Path::new(&cfg_path)) {
@@ -171,12 +169,12 @@ fn write_cfg(settings: &Value) -> Result<(), String> {
                 Err(why) => return Err(format!("Couldn't create cfg folder: {}", why)),
             }
         }
-    
-        let mut file = match File::create(format!("{}\\cfg\\melies.cfg", install_folder_str)) {
+
+        let mut file = match File::create(cfg_path.join("melies.cfg")) {
             Ok(file) => file,
             Err(why) => return Err(format!("Couldn't create melies.cfg: {}", why)),
         };
-    
+
         match file.write_all(cfg.as_bytes()) {
             Ok(_) => {}
             Err(why) => return Err(format!("Couldn't write melies.cfg: {}", why)),
@@ -248,7 +246,7 @@ fn start_vdm(vdm: &mut VDM, clip: &Clip, settings: &Value) {
             if let Some(skip_to_tick) = clip.start_tick.checked_sub(66) {
                 skip_props.skip_to_tick = Some(skip_to_tick);
             }
-            
+
             skip_props.name = format!("Skip to first clip");
         }
     }
@@ -365,7 +363,8 @@ fn record_clip(vdm: &mut VDM, clip: &Clip, settings: &Value) {
             "sparklyfx" => {
                 if settings["output"]["folder"].as_str().unwrap().len() > 0 {
                     commands = format!(
-                        "sf_recorder_start {}\\{}",
+                        "sf_recorder_start {}\\{}", // no need to change this as tf2 normalizes
+                                                    // paths internally
                         settings["output"]["folder"].as_str().unwrap(),
                         clip_name
                     );
@@ -436,7 +435,7 @@ fn check_dir(files: Result<fs::ReadDir, std::io::Error>) -> Result<String, Strin
         }
         Err(_) => {
             return Err(
-                "Could not find the _events.txt or KillStreaks.txt files.\r\nPlease check your settings to ensure the tf folder is correctly linked.\r\nIf you do not have either file, please make one in the \\tf or \\tf\\demos folder.".to_string()
+                "Could not find the _events.txt or KillStreaks.txt files.\nPlease check your settings to ensure the 'tf' folder is correctly linked.\nIf you do not have either file, please make one in the 'tf' or 'tf/demos' folder.".to_string()
             );
         }
     }
@@ -504,7 +503,7 @@ fn find_dir(settings: &Value) -> Result<String, String> {
     match settings["tf_folder"].as_str() {
         Some(tf_folder) =>
             Err(
-                format!("Could not find the _events.txt or KillStreaks.txt files.\r\nPlease check your settings to ensure the tf folder is correctly linked.\r\nIf you do not have either file, please make one in the \\tf or \\tf\\demos folder. \r\n\r\ntf_folder setting: ({})", tf_folder)
+                format!("Could not find the _events.txt or KillStreaks.txt files.\nPlease check your settings to ensure the tf folder is correctly linked.\nIf you do not have either file, please make one in the 'tf' or 'tf/demos' folder. \n\ntf_folder setting: ({})", tf_folder)
             ),
         None => Err("tf_folder setting not set".to_string()),
     }
@@ -534,7 +533,7 @@ fn ryukbot() -> Value {
         Err(_) => {
             return json!({
                 "code": 404,
-                "err_text": "Failed to read _events.txt or KillStreaks.txt\r\nPlease check your settings to ensure the tf folder is correctly linked.".to_string()
+                "err_text": "Failed to read _events.txt or KillStreaks.txt\nPlease check your settings to ensure the tf folder is correctly linked.".to_string()
             });
         }
     };
@@ -548,7 +547,7 @@ fn ryukbot() -> Value {
     if event_len == 0 {
         return json!({
             "code": 410,
-            "err_text": format!("_events.txt or KillStreaks.txt was found but had no valid events. Please add events before running again.\r\n\r\nFile Location: {}", dir)
+            "err_text": format!("_events.txt or KillStreaks.txt was found but had no valid events. Please add events before running again.\n\nFile Location: {}", dir)
         });
     }
 
@@ -770,7 +769,7 @@ fn save_events(new_events: Value) -> Value {
     let mut new_events_text = String::new();
 
     for demo in new_events.as_array().unwrap() {
-        extend!(new_events_text, "{}\r\n", ">");
+        extend!(new_events_text, "{}\n", ">");
 
         for event in demo.as_array().unwrap() {
             let re = Regex::new("\\[(.*)\\] (.*) \\(\"(.*)\" at (\\d*)\\)").unwrap();
@@ -790,14 +789,14 @@ fn save_events(new_events: Value) -> Value {
 
             if event["demo_name"].as_str().unwrap() != original_event.demo_name {
                 let built_event = build_event_from_json(event);
-                extend!(new_events_text, "{}\r\n", built_event.event);
+                extend!(new_events_text, "{}\n", built_event.event);
                 events.push(built_event);
                 continue;
             }
 
             if event["tick"].as_i64().unwrap() != original_event.tick {
                 let built_event = build_event_from_json(event);
-                extend!(new_events_text, "{}\r\n", built_event.event);
+                extend!(new_events_text, "{}\n", built_event.event);
                 events.push(built_event);
                 continue;
             }
@@ -806,14 +805,14 @@ fn save_events(new_events: Value) -> Value {
                 Bookmark(bm) => {
                     if event["isKillstreak"].as_bool().unwrap() {
                         let built_event = build_event_from_json(event);
-                        extend!(new_events_text, "{}\r\n", built_event.event);
+                        extend!(new_events_text, "{}\n", built_event.event);
                         events.push(built_event);
                         continue;
                     }
 
                     if bm.to_owned() != event["value"]["Bookmark"].as_str().unwrap() {
                         let built_event = build_event_from_json(event);
-                        extend!(new_events_text, "{}\r\n", built_event.event);
+                        extend!(new_events_text, "{}\n", built_event.event);
                         events.push(built_event);
                         continue;
                     }
@@ -821,21 +820,21 @@ fn save_events(new_events: Value) -> Value {
                 Killstreak(ks) => {
                     if !event["isKillstreak"].as_bool().unwrap() {
                         let built_event = build_event_from_json(event);
-                        extend!(new_events_text, "{}\r\n", built_event.event);
+                        extend!(new_events_text, "{}\n", built_event.event);
                         events.push(built_event);
                         continue;
                     }
 
                     if ks.to_owned() != event["value"]["Killstreak"].as_i64().unwrap() {
                         let built_event = build_event_from_json(event);
-                        extend!(new_events_text, "{}\r\n", built_event.event);
+                        extend!(new_events_text, "{}\n", built_event.event);
                         events.push(built_event);
                         continue;
                     }
                 }
             }
 
-            extend!(new_events_text, "{}\r\n", original_event.event);
+            extend!(new_events_text, "{}\n", original_event.event);
             events.push(original_event);
         }
     }
@@ -974,23 +973,11 @@ fn save_backup(settings: &Value) -> Value {
     let user_profile = env::var("USERPROFILE");
 
     let output_folder = match user_profile {
-        Ok(profile) => {
-            format!("{}\\Documents\\Melies\\backups", profile)
-        }
-        Err(_) => {
-            format!(
-                "{}\\backups",
-                std::env::current_exe()
-                    .unwrap()
-                    .parent()
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-            )
-        }
+        Ok(profile) => BACKUPS_DIR.join(profile),
+        Err(_) => BACKUPS_DIR.to_owned(),
     };
 
-    let output_path = format!("{}\\{}.txt", output_folder, date);
+    let output_path = output_folder.join(date);
 
     fs::create_dir_all(output_folder).unwrap();
 
@@ -1007,6 +994,9 @@ fn parse_log(url: Value) -> Value {
     parse(url)
 }
 
+const FIX_TF_FOLDER: &str =
+    "Can't find 'tf' folder. Please fix the \"'tf' Folder\" setting in settings.";
+
 #[command]
 fn load_vdms() -> Result<Value, String> {
     let settings = load_settings();
@@ -1014,9 +1004,7 @@ fn load_vdms() -> Result<Value, String> {
         return Ok(scan_for_vdms(settings));
     }
 
-    Err(String::from(
-        "Can't find \\tf folder. Please fix the \"\\tf Folder\" setting in settings.",
-    ))
+    Err(String::from(FIX_TF_FOLDER))
 }
 
 #[command]
@@ -1026,29 +1014,15 @@ fn load_demos() -> Result<Value, String> {
         return Ok(scan_for_demos(settings));
     }
 
-    Err(String::from(
-        "Can't find \\tf folder. Please fix the \"\\tf Folder\" setting in settings.",
-    ))
+    Err(String::from(FIX_TF_FOLDER))
 }
 
 pub(crate) fn validate_backups_folder() -> bool {
     let user_profile = env::var("USERPROFILE");
 
     let backups_folder = match user_profile {
-        Ok(profile) => {
-            format!("{}\\Documents\\Melies\\backups", profile)
-        }
-        Err(_) => {
-            format!(
-                "{}\\backups",
-                std::env::current_exe()
-                    .unwrap()
-                    .parent()
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-            )
-        }
+        Ok(profile) => BACKUPS_DIR.join(profile),
+        Err(_) => BACKUPS_DIR.to_owned(),
     };
 
     match fs::read_dir(backups_folder) {
@@ -1067,20 +1041,8 @@ pub(crate) fn scan_for_backups() -> Value {
     let user_profile = env::var("USERPROFILE");
 
     let backups_folder = match user_profile {
-        Ok(profile) => {
-            format!("{}\\Documents\\Melies\\backups", profile)
-        }
-        Err(_) => {
-            format!(
-                "{}\\backups",
-                std::env::current_exe()
-                    .unwrap()
-                    .parent()
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-            )
-        }
+        Ok(profile) => BACKUPS_DIR.join(profile),
+        Err(_) => BACKUPS_DIR.to_owned(),
     };
 
     for entry in fs::read_dir(backups_folder).unwrap() {
@@ -1121,23 +1083,11 @@ fn reload_backup(file_name: Value) -> Result<Value, String> {
     let user_profile = env::var("USERPROFILE");
 
     let backups_folder = match user_profile {
-        Ok(profile) => {
-            format!("{}\\Documents\\Melies\\backups", profile)
-        }
-        Err(_) => {
-            format!(
-                "{}\\backups",
-                std::env::current_exe()
-                    .unwrap()
-                    .parent()
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-            )
-        }
+        Ok(profile) => BACKUPS_DIR.join(profile),
+        Err(_) => BACKUPS_DIR.to_owned(),
     };
 
-    let file_path = format!("{}\\{}", backups_folder, file_name.as_str().unwrap());
+    let file_path = backups_folder.join(file_name.as_str().unwrap());
     let backup_file = Path::new(&file_path);
 
     let settings = load_settings();
@@ -1153,7 +1103,7 @@ fn reload_backup(file_name: Value) -> Result<Value, String> {
         }
     }
 
-    println!("Reloading backup: {}", file_path);
+    println!("Reloading backup: {}", file_path.display());
 
     if backup_file.exists() {
         let copy = fs::copy(file_path, dir);
@@ -1279,20 +1229,8 @@ fn load_theme() -> Value {
     let user_profile = env::var("USERPROFILE");
 
     let settings_path = match user_profile {
-        Ok(profile) => {
-            format!("{}\\Documents\\Melies\\theme.json", profile)
-        }
-        Err(_) => {
-            format!(
-                "{}\\theme.json",
-                std::env::current_exe()
-                    .unwrap()
-                    .parent()
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-            )
-        }
+        Ok(profile) => CONF_DIR.join(profile).join("theme.json"),
+        Err(_) => CONF_DIR.join("theme.json"),
     };
 
     if Path::new(&settings_path).exists() {
@@ -1314,40 +1252,20 @@ fn open_themes_folder() {
     let user_profile = env::var("USERPROFILE");
 
     let addons_path = match user_profile {
-        Ok(profile) => {
-            format!("{}\\Documents\\Melies", profile)
-        }
-        Err(_) => {
-            format!(
-                "{}",
-                std::env::current_exe()
-                    .unwrap()
-                    .parent()
-                    .unwrap()
-                    .to_str()
-                    .unwrap()
-            )
-        }
+        Ok(profile) => &*CONF_DIR.join(profile),
+        Err(_) => &*CONF_DIR,
     };
 
     fs::create_dir_all(&addons_path).unwrap();
 
-    Command::new("explorer").arg(addons_path).spawn().unwrap();
+    impl_open_file(&addons_path);
 }
 
 #[command]
 fn open_install_folder(install: &str) {
-    let settings = load_settings();
-    let tf_folder = settings["tf_folder"].as_str().unwrap();
-    let tf_parent = PathBuf::from(tf_folder).parent().unwrap().to_path_buf();
+    fs::create_dir_all(install).unwrap();
 
-    let relative_path = install.replace(&format!("{}\\", tf_parent.to_str().unwrap()), "");
-
-    let install_folder = format!("{}\\{}", tf_parent.to_str().unwrap(), relative_path);
-
-    fs::create_dir_all(&install_folder).unwrap();
-
-    Command::new("explorer").arg(install_folder).spawn().unwrap();
+    impl_open_file(install);
 }
 
 #[command]
@@ -1374,7 +1292,11 @@ fn cleanup_rename(demo_map: Value) {
     save_events(cleanup_renamed_events(demo_map.clone(), events));
     let settings = load_settings();
 
-    cleanup_renamed_vdms(demo_map, scan_for_vdms(load_settings()), settings["tf_folder"].as_str().unwrap());
+    cleanup_renamed_vdms(
+        demo_map,
+        scan_for_vdms(load_settings()),
+        settings["tf_folder"].as_str().unwrap(),
+    );
 }
 
 fn sanitize_name(path: &Path) -> PathBuf {
@@ -1435,11 +1357,11 @@ fn load_files(folder: &str) -> Value {
             continue;
         }
 
-        let has_layers = std::fs::read_dir(path.to_str().unwrap().to_string() + "\\take0000");
+        let has_layers = std::fs::read_dir(path.join("take0000"));
 
         let layers = match has_layers {
             Ok(layers) => layers,
-            Err(_) => continue
+            Err(_) => continue,
         };
 
         let mut video_layers = json!({});
@@ -1456,7 +1378,10 @@ fn load_files(folder: &str) -> Value {
                 continue;
             }
 
-            video_layers[file_name.replace(".mp4", "").replace(".avi", "").replace(".mkv", "")] = json!(path.to_str().unwrap().to_string());
+            video_layers[file_name
+                .replace(".mp4", "")
+                .replace(".avi", "")
+                .replace(".mkv", "")] = json!(path.to_str().unwrap().to_string());
         }
 
         let video: Value = json!({
@@ -1471,14 +1396,22 @@ fn load_files(folder: &str) -> Value {
     json!(videos)
 }
 
+fn impl_open_file(path: impl AsRef<OsStr>) {
+    opener::open(path).unwrap();
+}
+
+fn impl_delete_file(path: impl AsRef<Path>) {
+    trash::delete(path).unwrap();
+}
+
 #[tauri::command]
 fn open_file(path: &str) {
-    opener::open(path).unwrap();
+    impl_open_file(path)
 }
 
 #[tauri::command]
 fn delete_file(path: &str) {
-    trash::delete(path).unwrap();
+    impl_delete_file(path)
 }
 
 #[tauri::command]
@@ -1525,3 +1458,19 @@ fn main() {
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
 }
+
+static CONF_DIR: LazyLock<PathBuf> = LazyLock::new(|| {
+    if cfg!(windows) {
+        // Microsoft doesn't know what a standard is
+        home_dir().map(|d| d.join("Documents").join("Melies"))
+    } else {
+        // MacOS, Linux, *BSD, etc.
+        std::env::var_os("XDG_CONFIG_HOME")
+            .map(PathBuf::from)
+            .map(|d| d.join("Melies"))
+            .or_else(|| home_dir().map(|d| d.join(".config").join("Melies")))
+    }
+    .unwrap_or_else(|| current_dir().unwrap())
+});
+
+static BACKUPS_DIR: LazyLock<PathBuf> = LazyLock::new(|| CONF_DIR.join("backups"));
