+++ src/app.css
@@ -186,7 +186,9 @@ select {
   font-family: inherit;
   color: var(--pri-con-text);
   background-color: var(--bg2);
-  transition: border-color 0.25s, filter 0.3s;
+  transition:
+    border-color 0.25s,
+    filter 0.3s;
   outline: none;
   filter: drop-shadow(0px 0px -1px var(--pri));
 }
@@ -273,7 +275,9 @@ button:not(:disabled):hover {
   color: var(--pri-con-text);
   border-color: var(--pri-con);
   background-color: var(--bg2);
-  transition: border-color 0.25s, filter 0.3s;
+  transition:
+    border-color 0.25s,
+    filter 0.3s;
   outline: none;
   filter: drop-shadow(0px 0px -1px var(--pri));
 }
@@ -306,7 +310,9 @@ button:not(:disabled):hover {
   font-family: inherit;
   color: var(--err-con-text);
   border-color: var(--err-con);
-  transition: border-color 0.25s, filter 0.3s;
+  transition:
+    border-color 0.25s,
+    filter 0.3s;
   outline: none;
   filter: drop-shadow(0px 0px -1px var(--err));
 }
@@ -648,7 +654,7 @@ input:checked + .slider--tert:before {
 .autocomplete-input {
   height: 34px !important;
 }
-  
+
 .autocomplete-list {
   max-height: 200px !important;
   overflow-x: hidden;
@@ -656,7 +662,7 @@ input:checked + .slider--tert:before {
 
 .autocomplete-list-item {
   text-align: left;
-  padding-left: .5rem !important;
+  padding-left: 0.5rem !important;
   padding-right: 0 !important;
 }
 
@@ -703,8 +709,8 @@ input:checked + .slider--tert:before {
   border-left: 2px solid !important;
   border-bottom: 2px solid !important;
 
-  height: .35em !important;
-  width: .35em !important;
+  height: 0.35em !important;
+  width: 0.35em !important;
 }
 
 /* Track */
@@ -758,8 +764,8 @@ input:checked + .slider--tert:before {
   border-left: 2px solid !important;
   border-bottom: 2px solid !important;
 
-  height: .35em !important;
-  width: .35em !important;
+  height: 0.35em !important;
+  width: 0.35em !important;
 }
 
 /* Track */
