+++ src/lib/components/Switch.svelte
@@ -30,7 +30,7 @@
       class:tooltip--left={left}
       data-tooltip={tooltip?.trim() || ""}
       style={tooltip
-        ? `--kills: ${tooltip.split(/\r\n|\r|\n/).length - 1};`
+        ? `--kills: ${tooltip.split(/\n||\n/).length - 1};`
         : ""}
     >
       {title}
